Move ReadBuf and WriteBuf to io module
