+++ src/lib.rs
@@ -24,7 +24,6 @@ use std::sync::Arc;
 use std::time::Duration;
 use std::{fmt, ptr};
 
-pub mod buf;
 mod config;
 pub mod extract;
 pub mod fs;
